+++ app/routes/web.php
@@ -12,6 +12,7 @@ use App\Http\Controllers\Auth\SocialLoginController;
 use App\Http\Controllers\ChannelController;
 use App\Http\Controllers\CommentController;
 use App\Http\Controllers\HomeController;
+use App\Http\Controllers\ImageUploadController;
 use App\Http\Controllers\RssFeedController;
 use App\Http\Controllers\SubscriptionSettingsController;
 use Illuminate\Support\Facades\Route;
@@ -38,6 +39,9 @@ Route::get('/feeds/channels/{channel}.xml', [RssFeedController::class, 'channel'
 
 Route::middleware('auth')->group(function (): void {
     Route::post('/articles/{article}/comments', [CommentController::class, 'store'])->name('articles.comments.store');
+    Route::post('/uploads/images', [ImageUploadController::class, 'store'])
+        ->middleware('throttle:30,1')
+        ->name('uploads.images.store');
     Route::get('/settings/account', [AccountSettingsController::class, 'edit'])->name('settings.account.edit');
     Route::put('/settings/account/profile', [AccountSettingsController::class, 'updateProfile'])->name('settings.account.profile.update');
     Route::put('/settings/account/password', [AccountSettingsController::class, 'updatePassword'])->name('settings.account.password.update');